+++ inst/extdata/config.R
@@ -227,7 +227,7 @@ CONFIG$CERT_LINKS <- list(
   zenodo_api = "https://zenodo.org/api/records/",
   crossref_api = "https://api.crossref.org/works/",
   openalex_api = "https://api.openalex.org/works/",
-  researchequals_api = "https://www.researchequals.com/api/modules/main/"
+  researchequals_api = "https://researchequals.com/api/"
 )
 
 CONFIG$CERTS_URL_PREFIX <- "https://doi.org/"